+++ CHANGELOG.md
@@ -4,7 +4,7 @@ API is still not stable. Expect breaking changes until version 1.0.0.
 
 ## [Unreleased]
 ### Changed
-- Nothing yet
+- Changing at-least-one-container-ok for keep-some-containers-ok 
 
 ## 0.1.0 - 2016-10-22
 ### Added